(PC-9690) bookings: Raise error in `mark_as_used()` if already used
This is a revert of 9b2168054df98c64db4912b61e3af06516d958b3. In that
former commit, we made `mark_as_used()` a no-op if the booking was
already used. This was done to prevent errors when the offerer tried
to validate a booking that had been automatically marked as used.

This was both:

- useless: offerers never use our API to mark bookings as used for
  automatically-validated bookings. These bookings always have an
  activation code, which the beneficiary uses on the offerer's web
  site. The booking token is never used in that case.

- an error: for bookings that are not automatically validated upon
  creation, we (obviously) do NOT want to allow a token to be used
  twice. Therefore, the mark-as-used API must fail if it's given
  a booking that has already been used.