payments: Ignore free offers when looking for bookings to reimburse
It used to be implicit that free bookings were not reimbursed, because
"standard" rules yielded a 0 € reimbursement that was ignored (i.e. it
did not generate a Payment object).

However, an offer may now have a custom reimbursement rule. If one of
its stock used to be 0 € for a while (because of a manual entry
error), then the corresponding bookings would be free and yet we would
calculate a custom reimbursement amount (based on the real price of
the offer) for them. Of course we don't want to reimburse an offerer
for free bookings.

(Yes, it happened on stock 5160973 which had 4 free bookings: 601757,
615731, 624400 and 1226123. We blocked them manually, but that was
more work than we would like.)