(PC-6977) [FIX] booking confirmation email: missing departement code
If a user without any postal code information tried to book a digital
offer, this created an assertion error:

assert isinsance(departement_code, str)

The departement_code is used right after as a dict key. There is no need
to ensure that it is a str: any hashable object can be a dict key, and
None is a valid key dict.