(PC-9939) payments: Remove counter-productive error handling in payment generation script
If we fail too soon in the payment generation function, there is no
point in catching the exception and continuing. We will have to fix
the problems and re-execute the function. So let's not catch
exceptions and let them raise to stop the execution of the code.

Also, do not change the payment status on error. Two cases:

- when the generated XML is invalid, we should not update the payment
  status to `NOT_PROCESSABLE`. We would either have to run the full
  command again (including the call to `generate_payments()` that is
  very long, to mark those payments as `PENDING` again); or rollback
  those status changes before manually executing
  `generate_and_send_payments()` with the latest batch date.

  Now the payment statuses stay the same, and we can simply do the
  latter without having to touch the payment statuses.

- failing to send an e-mail should not change the payment status
  either. The files are stored on disk and can be sent manually.