(PC-10199) add venueTypeCode to Venue
We need to have a stable key to identify a Venue's type. We have a
VenueType model but it only has a label, which can change over time,
therefore we need to add a stable key so the front end client can map
icons/default image/etc to a given venue type.

To avoid huge changes at once, this commits proposes to add a new column
to Venue which will contain this code. The existing behaviour with the
VenueType.label column remains unchanged.

This is a first step change, this Venue.venueTypeCode / Venue.VenueType
situation is not meant to be remain.

Changes:

+ add venueTypeCode (enum as string) to Venue
+ return this new column on GET /venue/<venue_id>
+ add script that fills this column with the expected values, given a
  Venue's existing label (Venue.VenueType.label)