(PC-10444) Transform pydantic payload into json payload in cloud task decorator
It is useful to use the pydantic model both in the definition of the task endpoint body and in the argument of the task function. In RUNNING_TESTS mode, the type is then correct. However, when the endpoint is called for real, the pydantic model must be transformed into a dictionary