Fix PcObject.__eq__
The previous version of the code did not work when comparing an
instance of PcObject (or an instance of one its subclasses) with
anything that does not have an `id` object. Unfortunately, that's what
SQLAlchemy does in `fire_replace_event()`.

Also, the code was incorrect since it did not check that compared
objects were of the same type. Now it does.