(PC-9781) Fix wrong test
Since previous commit, when we chose a response status that is not listed by spectree.response.DEFAULT_CODE_DESC, the app cannot be instanciated because of spectree.response.__init__. Before previous commit, the app could be instanciated BUT spectree would throw an exception in response.generate_spec when trying to get swagger route. Since we don't use such codes in the app for the moment, just remove the test. We will adapt spectree.response if we need to use custom codes