(PC-10636) : add extra step for beneficiary only for newer clients
The addition of the extra steps indeed causes issues for native
applications that are nonnt able to understand the step and
redirects to the home.
This is an issue if the user uploaded its ID card but didn't
validate the general conditions in which case the user is locked
out of the beneficiary process.
The workaround is to act as usual for older native app and only
return the new extra step for newer versions.