(PC-11032) search: Use a meta-engine for offers
We used to have a single engine for all offers. Now we have a
meta-engine (named "offers-meta") that includes 6 engines (named
"offers-0" to "offers-5"). Each offer is indexed on its own engine,
depending on the offer id. Using such a system helps a lot in lowering
the response time of search requests.

Venues are still indexed in a single engine without any meta-engine
"above".