(PC-10591) index permanent venues only
+ rename async_index_venue_ids
to async_index_venues and changes its input: venues instead of venue
ids. Since the function needs to filter venues, and that each call to
the function already has a Venue object, it seems more logical to use
venues as input instead of taking ids and re-fetching venues from it.