(PC-7698) Turn clock-related entry points into commands
Each scheduled task "container" can now be called as a Flask command,
for example:

    $ flask titelive_clock

The goal of this commit is to make sure that we always load the
`pcapi.flask_app` module first, which means that we control the order
in which modules are imported. It should help avoid import loops.

Also, `install_logging()` is removed because it's already done when
the application is loaded from `pcapi.flask_app`.