(PC-7698) Turn worker entry point into a command
The worker can now be called as a Flask command:

    $ flask worker

The goal of this commit is to make sure that we always load the
`pcapi.flask_app` module first, which means that we control the order
in which modules are imported. It should help avoid import loops.

Also, `install_logging()` is removed because it's already done when
the application is loaded from `pcapi.flask_app`.