(PC-4708) requirements: Stop using nltk and hard-code stopwords instead
We were installing the `nltk` third-party library to get a list of
stopwords. This forced us to download a stopwords file (and another
file for the punkt tokenizer, for a reason unknown to me). Since
f9aabd2c2b, we have a hard-coded list of stopwords, which is unlikely
to evolve anyway.