(PC-11330) feat:account: update email
Update POST /profile (native v1): add two input parameters: email and
password. They are both needed to start the email update process. The
first one is the new email adress and the password is a security option.

The email is not updated immediately: an email with a validation
link is sent to this new email adress.

A new email address is invalid if an existing email exists.

*email validation*

Add API route (native) to validate email, using a JWT token.
It is almost identical to the webapp version.

+ update change_email in order to move some validations to a
  pydantic model which will make the code smaller and more readable.

*save TTL*

Save the token's TTL since we are using a JWT that is not saved into the
database. This prevents from having two (or more) active token for the
same user. And we will be able to send this information to the front end
(needed for an incoming feature).

Token TTL is saved using Redis: is not a vital information and redis'
tools like incr makes things easier.