search: Turn full_index_offers into a Flask command and support Algolia
The script was written to index offers on App Search but it's now
useful on Algolia as well and could be useful in the future. It now
supports both App Search and Algolia, and it's a command:

    $ flask full_index_offers algolia 1 10_000_000

Also, fix bug that caused reporting to happen every 11_000 items
instead of every 10_000 items. And use new `is_eligible_for_search`
property.