(PC-11229) Update get_wallet_balance and check_booking function
Now the balance of a deposit is computed with the bookings attached to it. The wallet_balance is the deposit of the not expired deposit or is 0. If a former deposit's bookings is modified, we need to check the balance of this speicific deposit, so we add a get_deposit_balance funcion