(PC-11701) : fix in-app message crash for IDCheck's unread-document
The in-app message for IDCheck's unread-document error possibly
has a race issue that could result in a crash and retry whenever
a user had the unread-document code after it confirmed the
general conditions.
General conditions would flag the user as hasCompletedIdCheck
which would raise an exception when requested a new idcheck
token leading to a post IDCheck crash and endless retries
on IDCheck step.