api: More compact output of pre-commit checks
Now each blocking check shows a line like:

    Running isort to organize imports...

When the check is done, the line is *replaced* by a status line, for
example:

    isort: Imports correctly sorted

If the checker reported warnings, you get a nice output, such as:

    black: Add edited files to your commit:
      - api/path/to/file.py
      - api/path/to/other/file.py

This commit also removes a bunch of gratuitous newline characters. It
makes for a much more compact output.