api: hooks: Fix Alembic pre-commit check
Git hooks are run from the root of the Git repository. As such,
running `alembic` fails because it cannot find its configuration
file. We could provide it (`alembic --config api/alembic.ini`). But
this file mentions `src/pcapi/alembic/` as the path of the "alembic"
folder, which cannot be found from the root of the Git repository. We
could specify `%(here)s/src/pcapi/alembic` and Alembic would find the
directory... but there would be an error while importing modules
because required environment variables are missing, since `.env.` file
are in the `api/` directory (and thus not read when running from the
Git repository).