[PC-11985][API] Improve search performance on pro's offer page
Two parts were required to improve the search.
First a missing index on extraData -> isbn forced the query to scan
the entire table which takes too much time.
Second, searching for the name OR the ibsn made Postgres scan the
table instead of using the indexes. By using an union instead
of a OR condition, Postgres can spawn two worker to search against
the indexes and then create the result set.