[PC-12017][api] Do not import models from modules that define models
In a module that defines a model, we should not import other models or
modules that directly or indirectly import models (such as
`pcapi.xxx.api` modules) because we risk import loops.

I left a few occurrences because some import were harmless (for now,
at least...) and/or replacing them would have made the code too ugly.

This is the second part of the "No more import loop" series.