[PC-12506][API] Fix OOM on scheduled tasks
Reworked a bit the way user automation scheduled tasks were
handled.
A user subscription workflow has recently been introduced.
That workflow is initiated every time a User object is created
or loaded. During its initialization, it requires
`subscriptionState` and `phoneValidationStatus` fields to be loaded.

User automation queries only load the email field. This results in
2 extra queries for each user to get the `subscriptionState` and
`phoneValidationStatus` fields.

In order to work around this and since it really only needs the
email, this commits replaces the use of the User model with
simple email strings which does not trigger the workflow
events any longer.