(PC-12483)[api]: correct wrong status in csv that lists bookings
For eac bookings, the status is "confirme" only when
cancellation_limit_date has passed and the booking has been confirmed by
the principal. In the case where the cancellation_limit_date
HAS NOT PASSED and the booking has been confirmed by the principal, the status is "reserve". Finally, if the booking has not been
confirmed by the principal, the status is "prereserve" be it before or
after cancellation_limit_date (excluding CANCELLED, USED and REIMBURSED
status which are always what they say they are)