[bsr][api][tests] Keep ReferenceScheme in the test database
There is a migration that creates multiple ReferenceScheme. We can use
that as data fixtures for tests, to avoid having to create a
ReferenceScheme each time a test generates invoices. For this to work,
we must not clean the table in `clean_database`, adapt the factory to
reuse existing instances, and be a bit more careful in tests that do
need to use `ReferenceSchemeFactory`.