[PC-13231][api][offers] Allow to change stock price if offer custom rule is not valid anymore
When creating a stock or editing the price of an existing stock, we
used to check if the offer had a custom reimbursement rule. If so, we
would reject the request.

Now we allow the creation or edition if the reimbursement rule is in
the past (i.e. if it's not active anymore). If the rule is active or
will be active, we still refuse to create or modify a stock.