[PC-13265][API] use Spectree's Response in spectree_serialize
Remove the former `code_descriptions` from the `spectree_serialize`
and replace it with raw Spectree's Response. This should make it
possible to describe the associated response whenever an HTTP
error is returned and will provide more informations to the client
about what to expect when things go south.
Also fix some typing.