[PC-13823][api][providers] Properly catch provider synchronization exception
An SQL integrity error in `provider.updateObjects()` was improperly
caught by the `do_update()` wrapper, which caused an error on any
consecutive SQL query. In practice, the "Allociné" synchronization
failed and stopped half-way on the first SQL query after an SQL
integrity error.

This is a very minimal patch (just enough for a hotfix) but more needs
to be done in this module to clean it up (and clean its tests that
overly use mocks).